requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
imageio==2.36.0
lxml==5.3.0
numpy==2.1.3
pillow==11.0.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
